+++ sysspecter/__init__.py
@@ -3,4 +3,4 @@
 See everything. Find the cause.
 """
 
-__version__ = "1.3.0"
+__version__ = "1.3.1"